Revise architecture section in README
Going to be overhauling an old project and really making the product versatile across the desktop, mobile, and website views. Currently, I'll be trying to reshape the backend views and instead attach to the Google Cloud Platform for support. 

Again, this diagram is not exhaustive and many new features will change this diagram. For now, a high level view of what all features are coming will show here.